Add files via upload
This is my initial firmware for upgrading the B1 SE Plus to TMC2226 (or TMC2209) as they are the same in the code. 
Please note: This firmware will only drive on extra fan header on the motherboard, so you will want a Y adapter to run both. Also, this fan header will only be active when the stepper divers are active. After they turn off, the fan header will be active for 60secs and then shut off the fans.